Add separate static jinja script for resume
The full story is this:

https://en.wikipedia.org/wiki/Wikipedia:Emerson_and_Wilde_on_consistency

The foolish consistency for me was that the resume was being processed
with everything else. It's a foolish consistency because if you think
about it, the confines of a resume aren't really the same as most other
general purpose writing. I wrestled with this mistake for about a year
before I realized my mistake. Today I fixed it, and while the fix isn't
perfect, it's dramatically better than what I had before.

I'll make a few tweaks in the coming days, but for now, I just needed
to watch that the main script didn't delete the whole thing!